renaming some files, formatting
